Added explicit cast to the return value of strlen() in pico_protocol.c.
The cast tackles a warning only found when you compile the code for a
64-bit machine (64-bit size_t to uint32_t)

+++ stack/pico_protocol.c
@@ -335,7 +335,7 @@ void pico_protocol_init(struct pico_protocol *p)
     if (!p)
         return;
 
-    p->hash = pico_hash(p->name, strlen(p->name));
+    p->hash = pico_hash(p->name, (uint32_t)strlen(p->name));
     switch (p->layer) {
     case PICO_LAYER_DATALINK:
         pico_tree_insert(&Datalink_proto_tree, p);
